controllers

+++ demo/src/main/java/FaceSale/demo/api/repositories/UsuarioRepository.java
@@ -2,5 +2,5 @@ package FaceSale.demo.api.repositories;
 
 import FaceSale.demo.api.entities.Usuario;
 
-public interface UsuarioRepository extends BaseRepository<Usuario, Integer> {
+public interface UsuarioRepository extends BaseRepository<Usuario, Long> {
 }
